Include b2b inventory display e2e
