+++ projects/storefrontapp-e2e-cypress/cypress/helpers/inventory-display.ts
@@ -1,3 +1,4 @@
+import { visitProductPage } from '../helpers/coupons/cart-coupon';
 import * as sampleData from '../sample-data/inventory-display';
 
 export const stockSelector = 'cx-add-to-cart .info';
@@ -72,6 +73,65 @@ export function assertInventoryDisplay(productCode: string, alias: string) {
       } else {
         expect(text).to.equal(`${sampleData.stockLabel}`);
       }
+    });
+  });
+}
+
+export function testInventoryDisplay(productCode: string) {
+  const productDetailsAlias = interceptProductDetails(productCode);
+  visitProductPage(productCode);
+
+  cy.wait(`@${productDetailsAlias}`)
+    .its('response.statusCode')
+    .should('eq', 200);
+
+  assertInventoryDisplay(productCode, `@${productDetailsAlias}`);
+}
+
+export function runInventoryDisplayE2E(consumer: string, sampleData: any) {
+  context(`${consumer} - Inventory Display`, () => {
+    before(() => {
+      cy.window().then((win) => win.sessionStorage.clear());
+    });
+
+    describe('Inventory Display - disabled', () => {
+      before(() => {
+        configureInventoryDisplay(false);
+      });
+
+      it('should NOT render number of available stock', () => {
+        testInventoryDisplay(sampleData.IN_STOCK_WITH_QUANTITY_PRODUCT);
+      });
+    });
+
+    describe('Inventory Display - active', () => {
+      beforeEach(() => {
+        configureInventoryDisplay(true);
+      });
+
+      it('should render number of available stock', () => {
+        testInventoryDisplay(sampleData.IN_STOCK_WITH_QUANTITY_PRODUCT);
+      });
+
+      it("should render 'out of stock' if stock level 0 and inventory display is on", () => {
+        testInventoryDisplay(sampleData.OUT_OF_STOCK_PRODUCT);
+      });
+
+      it("should render 'In Stock' if force inStock status and inventory display is on", () => {
+        testInventoryDisplay(sampleData.FORCE_IN_STOCK_PRODUCT);
+      });
+
+      it('should render + if threshold applied and inventory display is on', () => {
+        testInventoryDisplay(sampleData.THRESHOLD_STOCK);
+      });
+
+      it('should NOT render + if threshold greater than stock level and inventory display is on', () => {
+        testInventoryDisplay(sampleData.STOCK_LESS_THAN_THRESHOLD);
+      });
+
+      it('should NOT render + if threshold equal to stock level and inventory display is on', () => {
+        testInventoryDisplay(sampleData.STOCK_EQUAL_THRESHOLD);
+      });
     });
   });
 }
