+++ projects/storefrontapp-e2e-cypress/cypress/integration/b2b/regression/inventory-display/inventory-display.e2e-spec.ts
@@ -1,0 +1,4 @@
+import { runInventoryDisplayE2E } from '../../../../helpers/inventory-display';
+import { inventoryDisplayB2B } from '../../../../sample-data/inventory-display';
+
+runInventoryDisplayE2E('B2B', inventoryDisplayB2B);
