+++ projects/storefrontapp-e2e-cypress/cypress/sample-data/inventory-display.ts
@@ -1,10 +1,22 @@
-export const IN_STOCK_WITH_QUANTITY_PRODUCT = '358639';
-export const OUT_OF_STOCK_PRODUCT = '872912';
-export const FORCE_IN_STOCK_PRODUCT = '1377492';
+export const inventoryDisplayB2C = {
+  IN_STOCK_WITH_QUANTITY_PRODUCT: '358639',
+  OUT_OF_STOCK_PRODUCT: '872912',
+  FORCE_IN_STOCK_PRODUCT: '1377492',
 
-export const THRESHOLD_STOCK = '280916';
-export const STOCK_LESS_THAN_THRESHOLD = '479742';
-export const STOCK_EQUAL_THRESHOLD = '458542';
+  THRESHOLD_STOCK: '280916',
+  STOCK_LESS_THAN_THRESHOLD: '479742',
+  STOCK_EQUAL_THRESHOLD: '458542',
+};
+
+export const inventoryDisplayB2B = {
+  IN_STOCK_WITH_QUANTITY_PRODUCT: '2222485',
+  OUT_OF_STOCK_PRODUCT: '1979039',
+  FORCE_IN_STOCK_PRODUCT: '3788618',
+
+  THRESHOLD_STOCK: '693923',
+  STOCK_LESS_THAN_THRESHOLD: '2217258',
+  STOCK_EQUAL_THRESHOLD: '1128762',
+};
 
 export const stockLabel = 'In stock';
 export const stockOutOfStockLabel = 'Out of stock';
